+++ .gitignore
@@ -1,4 +1,5 @@
 *.out
+*.o
 .vscode/
 1dim_fem.dSYM/
 1dim_fem
