+++ define.h
@@ -1,0 +1,2 @@
+#pragma once
+#define N 1000
